feat: check rimraf call

+++ packages/electron/src/ElectronFileAccess.ts
@@ -205,7 +205,12 @@ export class ElectronFileAccess implements INativeFileAccess {
     public async deleteDirectory(path: string, storage: NativeFileStorage = NativeFileStorage.Data): Promise<Result<void>> {
         try {
             const directoryPath = this.getStoragePath(path, storage);
-            // TODO: JSSNMSHDD-2686 (Check this rimraf call!)
+
+            if (!directoryPath.includes(__dirname)) return Result.fail(this.resolveErrorCode("EACCES"));
+            if (!directoryPath.includes(this.datadir)) return Result.fail(this.resolveErrorCode("EACCES"));
+            if (!directoryPath.includes(this.homedir)) return Result.fail(this.resolveErrorCode("EACCES"));
+            if (!directoryPath.includes(this.tmpdir)) return Result.fail(this.resolveErrorCode("EACCES"));
+
             await new Promise((resolve) => rimraf(directoryPath, resolve));
             return Result.ok(undefined);
         } catch (err) {
